Made some changes to MiniMaxSearchTree.
Problem now is that it is not possible to add more that 1 set of children. The code executes but there still is only 1 set of children. In order to have more than 1 set of children then the code needs to be recurcive and new keyvalues have to be sent to the mehtod add(Node<K,V>, K, V);

The game proceeds which can be seen in the output and the AI agent plays with it self random moves.